Move 2nd tree selector to Tree Options
