+++ src/components/controls/choose-dataset.js
@@ -1,12 +1,7 @@
 import React from "react";
-import Select from "react-select";
 import { connect } from "react-redux";
 import ChooseDatasetSelect from "./choose-dataset-select";
 import parseParams from "../../util/parseParams";
-import { SelectLabel } from "../framework/select-label";
-import { loadTreeToo } from "../../actions/loadData";
-import { controlsWidth } from "../../util/globals";
-import { REMOVE_TREE_TOO } from "../../actions/types";
 
 // // remove starting or trailing slashes from path
 // const tidyUpPathname = (pathname) => {
@@ -43,7 +38,6 @@ class ChooseDataset extends React.Component {
     selectors array will only have 1 element */
     const selectors = []; // list to contain the different data set selectors
     let level = this.props.availableDatasets; // pointer used to move through the hierarchy -- currently at the top level of datasets
-    let treeToo;
     for (let vi = 0; vi < fields.length; vi++) {
       if (choices[vi]) {
         // pull options from the current level of the dataset hierarchy, ignore 'default'
@@ -58,48 +52,9 @@ class ChooseDataset extends React.Component {
             />
           </div>
         ));
-        if (Object.keys(level).indexOf("segment") !== -1 && options.length > 1) {
-          const treeTooOptions = options.filter((v) => v !== choices[vi]);
-          if (this.props.showTreeToo) {
-            treeTooOptions.unshift("REMOVE");
-          }
-          treeToo = {
-            fieldIdx: vi,
-            options: treeTooOptions,
-            treeOne: choices[vi] // the option chosen for tree one (e.g. NA, PB1...)
-          };
-        }
         // move to the next level in the data set hierarchy
         level = level[fields[vi]][choices[vi]];
       }
-    }
-
-    /* second tree? */
-    if (treeToo) {
-      selectors.push((
-        <SelectLabel key="treetootitle" text="Second Tree"/>
-      ));
-      selectors.push((
-        <div key={"treetooselect"} style={{width: controlsWidth, fontSize: 14}}>
-          <Select
-            name="selectTreeToo"
-            id="selectTreeToo"
-            value={this.props.showTreeToo}
-            options={treeToo.options.map((opt) => ({value: opt, label: opt}))}
-            clearable={false}
-            multi={false}
-            onChange={(opt) => {
-              if (opt.value === "REMOVE") {
-                this.props.dispatch({type: REMOVE_TREE_TOO});
-              } else {
-                const dataPath = [...choices];
-                dataPath.splice(treeToo.fieldIdx, 1, opt.value);
-                this.props.dispatch(loadTreeToo(opt.value, dataPath.join("_")));
-              }
-            }}
-          />
-        </div>
-      ));
     }
 
     // return a list of selectors in the order of the data set hierarchy
